Resistencia añadida y se reescala el tamaño del fuego

+++ GameofSorts/src/Interface/Board.java
@@ -26,10 +26,12 @@ public class Board extends JPanel implements ActionListener, MouseListener{
 	static SimpleLinkedList<Dragon> dragonsArray;
 	static SimpleLinkedList<FireBall> fireballs;
 	
+	private int cont;
 	
 	public Board() {
 		p = new Player(); //jugador
 		createWave();
+		cont = 1;
 		
 		fireballs = p.getFireballs();
 
@@ -109,9 +111,15 @@ public class Board extends JPanel implements ActionListener, MouseListener{
 			for(int j=0; j < fireballs.getLength(); j++) {
 				FireBall ftemp = fireballs.get(j).getData();
 				if(dtemp.getDragonHitBox().collidesWith(ftemp.getFireHitBox())) {
-					dtemp.alive = false;
-					dragonsArray.deleteByElement(dtemp);
-					ftemp.visible = false;
+					if(dtemp.getResistance() == cont) {
+						dtemp.alive = false;
+						dragonsArray.deleteByElement(dtemp);
+						ftemp.visible = false;
+						cont = 1;
+					}else {
+						ftemp.visible = false;
+						cont++;
+					}
 				}
 			}
 		}
